small fixes and restructuring and removed dask deps #46
