+++ pyproject.toml
@@ -10,7 +10,6 @@ authors = [
   { name = "Hessel Winsemius", email = "winsemius@rainbowsensing.com" }
 ]
 dependencies = [
-    "dask",
     "matplotlib",
     "numpy",
     "numba",
