+++ requirements.txt
@@ -1,4 +1,3 @@
-dask
 matplotlib
 numpy
 numba
